Bump memory for Celery Beat to 512M
This fixes the app crashing - apparently due to a lack of free
memory. Looking over the last month, we can see it's been hovering
at ~90% utilisation. 512M seems like a good compromise to avoid
this happening again vs paying for what we use.

The limit was last changed 5 years ago:

https://github.com/alphagov/notifications-api/pull/882