Refactor yearly usage API into functions per type
This makes it easier to extend each function with costs and free
allowances - especially for SMS.

In each function I've started using the "chargeable" terminology,
which we should eventually change in the API.

I've chosen to duplicate the "WHERE" clause in each subquery vs.
the top-level query. This will make more sense in later commits
where we start adding free allowance calculations, which need to
be done on a yearly basis - knowledge the subqueries should have.