Use admin_request consistently in billing tests
This avoids a bunch of boilerplate and makes it easier to see what
is being passed in the request.

Note that the "invalid_schema" test is slightly different because
it was previously passing an invalid JSON object - "{}" - instead
of a valid JSON but invalid by the schema - "\"{}\"". The new test
seems more valuable than the old one.