Add endpoint to archive an organisation
The new endpoint checks that an org has no users, pending invited users,
or services (in any state). If it does, admin will display an error message
so that the platform admin user archiving the org can manually reassign
or remove the team members and services.

Assuming that the org can be archived, we remove its email branding,
letter branding and domains, set its state to `active=False`, and change
its name so that the original name can be reused by new organisations.