create nested subquery to separately sum FY vs start/end range values
we need to aggregate `free_allowance_used_to_date` over the entire
financial year.

we need to aggregate `sms_cost`, `chargeable_sms`, etc over the range
provided in start_date and end_date

if we just try and filter the subquery on start and end, we'll exclude
those rows earlier in the financial year. This is fine except in the
edge case where a service has ft billing rows before the start date, but
none in teh start date. then, the filter will exclude those, there's no
rows within the range that we can get an up-to-date
`free_allowance_used_to_date` from, and we can't calculate sms_remainder
correctly.

the least-bad solution that we could come up with is to filter on the
entire financial year, but then when summing up sms_cost etc, instead of
just filtering the entire subquery, using a CASE statement inside the
SUM to only sum up those rows for that particular aggregate.