Remove redundant cases in base org report test
There quite a lot going on here so let's break it down:

- We were calling set_up_usage_data to give us a bunch of services
in orgs, with data in ft_billing:

  - org_1:
    - service_1_sms_and_letter
    - service_with_out_ft_billing_this_year
  - org_2:
    - service_with_emails
  - no org:
    - service_with_sms_within_allowance
    - service_with_letters_without_org
    - service_with_sms_without_org
  - org_for_service_with_letters:
    - service_with_letters

- We were also creating an extra "service_with_emails_for_org" and
adding some billing data for that, since "org_1" doesn't have any
services that sent email.

This change switches to using simpler input data. Comparing with the
old data, we can verify we don't lose any coverage:

1. There were 3 results. This implicitly checked that we exclude the
services in other orgs. This is already covered explicitly [^1].

2. We had results about SMS and letters and (separately) emails. In
the new test these are covered together in a single result.

The new test data from set_up_yearly_data also includes a few rows
before and after the billing period we're using in the test [^2].
This means we're covering the "only_queries_present_year" test and
it's safe to delete it.

[^1]: https://github.com/alphagov/notifications-api/blob/c4c9171cbaf27facb9f989e91d9741a94f19ce2e/tests/app/dao/test_fact_billing_dao.py#L942
[^2]: https://github.com/alphagov/notifications-api/blob/c4c9171cbaf27facb9f989e91d9741a94f19ce2e/tests/app/dao/test_fact_billing_dao.py#L43