Lift up test for mutli-service org SMS usage
This was testing the query directly, which made it inconsistent
with all the other tests. By using the test data defined in this
file we can significantly simplify the test setup.

I had to manually calculate the new numbers but the test is still
checking the same thing: that we correctly calculate the stats for
each service, individually. Although the test data is the same, we
can use different free allowances to prove the calculations are
performed separately, which is the aim of the test.

In order to reuse the test data for multiple services I had to
pass in a unique service name - otherwise we get a DB constraint
error for the "email_from" column, which is calculated based on
the name - for test data [^1].

[^1]: https://github.com/alphagov/notifications-api/blob/403744ef11624a4e9bf72aeb0672f5732d62a9e8/tests/app/db.py#L143