Push org usage coalesces into the query function
This is a step towards calculating the free_allowance_left in the
subquery by ensuring the data that feeds into it is not null. Both
coalesces are covered by

    test_fetch_usage_for_organisation_no_usage

and the outerjoin is required by

    test_fetch_usage_for_organisation_without_annual_billing

Moving the calculation to the subquery will support reusing it for
the all service SMS usage function in later commits.