Support variable rates for all service SMS usage
There are two reasons why the old code did not support multiple
SMS rates in the same financial year:

- The query contained a group by clause on the rate column, which
means we would get multiple rows if there are multiple rates, even
though the calling code only expects one.

- The query needs to go through ft_billing rows in date order when
attributing any remaining free allowance, as later rows may have a
diffeent cost if the rate changed in the period.

In this commit we reuse the _query function for orgs to calculate
usage, since it supports multiple rates.

What's changed
==============

The diff is quite big so let's break it down:

- The new test for variable rates is basically a duplicate of the
one we have for organisations [^1]. Using the new _query function
means we get support for variables SMS rates.

- The new _query function is a copy of the one we have for orgs,
just without the org parameter / filter. The way we're using the
query function at the top-level is the same.

The _query function is returning rows from ft_billing, annotated
with some extra columns like "charged_units" that take previous
usage into account, as well as the free allowance for the service.
The top-level function then aggregates the rows by service.

Potential gotcha
================

In our first attempt at this change the filter clause only had a
condition on the end date [^2], as we were concerned the subquery
would be affected by a condition on the start date and ignore any
previous spending. We used a helper function instead [^3].

The approach here is simpler and works. I'm not sure how we got to
the original approach, but it may be related to the mixed use of
"coalesce", which we sorted out in previous commits.

Conceptually, all we're doing is using the subquery to return all
the rows for the year, which are individually all correct; we then
use the "filter" clause to drop all rows except the (still correct)
ones for the specified period, which we then aggregate.

Review pointers
===============

In case it helps your review, here is some more detail about the
individual lines that are changing in this commit:

- Since we're no longer referring to FactBilling at the top-level,
we can remove the join and group by column - the subquery does the
same SMS-only filter and grouping by rate is no longer necessary
as the subquery looks at ft_billing rows individually.

- We're using the same min / max approach for the SMS remainder
and free allowance as we do for orgs. Because we're using aggregate
functions we no longer need the old group by clauses.

[^1]: https://github.com/alphagov/notifications-api/blob/7c592cae4933bab6267d130dd5df332ed5694ba7/tests/app/dao/test_fact_billing_dao.py#L869
[^2]: https://github.com/alphagov/notifications-api/pull/3548/files#diff-8e5521edad2738f137b117c5cec89497247a508a30bd21065dfe7e8d44aa8dddR97
[^3]: https://github.com/alphagov/notifications-api/pull/3548/files#diff-8e5521edad2738f137b117c5cec89497247a508a30bd21065dfe7e8d44aa8dddR53