Add missing tests for inactive services
This was only covered implicitly in one of the API-level tests [^1].
We populate annual billing when we create a service [^2] and then
manually before the start of each financial year [^3], but not for
inactive services. Ideally these functions should also produce a
sensible result when there's no annual billing anyway.

[^1]: https://github.com/alphagov/notifications-api/blob/ab9c2be813aa885bef201fded6012923dd050fd2/tests/app/organisation/test_rest.py#L942
[^2]: https://github.com/alphagov/notifications-api/blob/ab9c2be813aa885bef201fded6012923dd050fd2/app/service/rest.py#L263
[^3]: https://github.com/alphagov/notifications-api/blob/ab9c2be813aa885bef201fded6012923dd050fd2/app/commands.py#L783