Standardise free SMS allowance for base case tests
This will make it possible to DRY-up the test setup consistently.
I chose an allowance of 1 so we get a bit more value out of the
base case tests by making more precise assertions on more fields.

Since we no longer have a positive sms_remainder for the org and
all service functions, I've added a couple of explicit test cases
to cover that scenario - it's only covered incidentally elsewhere.