Make yearly billing setup function a fixture
This is consistent with how we write most tests. I've used a more
specific name to discourage future extension. While it's not ideal
to have the setup far away from each test, using this fixture:

- Avoids extensive setup in the 4 tests that use it, which detracts
from the assertions at the end.

- Reduces the mental load on someone working across these 4 tests -
they only need to understand the setup once.

I've moved the creation of FY 2016 free allowance to the fixture,
so all the numbers are in one place and to be consistent with the
other free allowances for 2015 and 2017.