Add Docker hostname to access host machine network
By default docker containers are meant to have access to the host machine
network via host.docker.internal name which points to the host network
gateway IP. This allows access to services running on the host machine
from inside the container.

This appears to have stopped working on Docker for Mac v20.10.13 running
on M1.

We can use the `--add-host` directive to manually add the host pointing
to the host-gateway IP.

Signed-off-by: Richard Baker <richard.baker@digital.cabinet-office.gov.uk>