Add deployment script
This script uses the CF API in order to:

1. Copy a droplet to an app
2. Trigger a deployment using this specific droplet
3. Monitor the deployment status for up to 15 minutes

It requires the files `./git/short_ref` and `api-droplet-guid-*.txt`
to be present in the root directory

The reason for doing this via the API is because the CLI doesn't provide
a way to copy the droplet between spaces (we need to download it and
upload it again) and it also doesn't support `cf push` providing just a
droplet guid (we need to provide a path to the droplet package on the
filesystem).

Given that the point of this is to improve deployment times,
it's unnecessary and wasteful to download and upload 100s of MBs
for all the apps.