Make deployment script a bit more defensive
The state of the world is different depending on whether we run it from
our laptops or from concourse.

Concourse doesn't have a git repo and only has one droplet-guid file,
whereas locally we have a git dir but we may have multiple droplet-guid
files, so we need to pick one.