refactor the org billing function
currently it sets up a bunch of data and then overwrites it from the
results of the sms billing - there's no reason for that billing query to
be the thing that records whether the service is active, especially
since if we merge it with the platform admin query it won't return rows
that have no billing data.

Pulling that stuff out makes the code a little easier, although it did
require creating a new function to return the services along with their
allowance for a given year.