add notify_user to api_client_request
just ensures we're always tearing down any db. in reality this isn't
required - we might have a test that doesn't need the DB at all to
verify that it's working, but the number of those tests should be so
small that any performance gain by not creating and deleting hte db
should be minimal

notify_user is useful because that user is required to be in the
database for precompiled templates to be created.