Add endpoint to update email branding pool
This adds an endpoint to update the email branding pool for an
organisation. The form that the data comes from allows multiple email
brandings to be added at once, so the endpoint takes this into account.
I've added a new dao function which allows bulk adding to the email branding
pool - this saves having to iterate through each new branding and doing
a database update for each one.

As with other POST endpoints, we check the data using a JSON schema,
however the request from admin should always have the data in the
required format, so I would not expect this to raise a validation error
in production. Instead, it acts to document the endpoint and to ensure
any future changes don't break things unexpectedly.