Add email verification service permission
While we're building out the document download email verification flow
we want to make sure the features are unavailable to most services, so
we'll use a service permission to essentially feature flag it as
disabled for everyone.

This gives us the ability to gradually allow services to test the new
features if we want to do a limited roll-out initially.