Add verify_email param to personalisation
Adds a new supported parameter to file uploads in email personalisation
- verify_email_before_download. This will be used to trigger the
verification email flow on the document download frontend. If this
parameter is present, we include `verification_email=email` in the call
to document download, which will result in storing a hashed email
against the s3 object. The presence of this hashed email will send the
user down the new verification flow.