Add validation for send-a-file upload dicts
If a user sends the wrong kind of data/value for the new
send-a-file/document download security features, let's provide a
user-friendly error message.

I would've ideally liked to do this as part of the generic incoming
notification JSON schema, but in order to do that we'd need to use a
`oneOf` subschema to allow for multiple data types and different
validation depending on the vlaue of personalisation fields. When we do
this, we lose the ability to (cleanly) provide explicit error messages
about what exactly is wrong with the input, because JSON Schema wraps
the root error message in some contextual one about all of the
sub-schemas failing to validate.

So we've decided to just do the validation on known-file upload
personalisation, which allows us to more cleanly present the error to
the user.