Optional: add personalisation schema
I think we broadly only expect simple datatypes (strings, maybe numbers)
as values in the personalisation dict. If it's a dict, we should expect
that it's a file upload and enforce the parameters we expect in this
case.

We might not want to accept this if we think that users might be sending
dicts for other reasons. This probably _shouldn't_ be happening - they
should serialize it to a string first. But it's possible they don't -
and we might want to not break their integrations if that's the case.