Force BeautifulSoup to use utf-8 encoding
If not explicitly told what the encoding of the data is, BeautifulSoup
tries a few different ways to guess what it is, using things like
cchardet/charset_normalizer. On one of our "broadcast alert content too
long" tests, this returns iso_8859_16 for the data.

This character encoding is known on most developer machines and so the
text is treated as being this encoding. However, that is actually the
wrong encoding - it's actually a utf-8 string.

This test passes on CI because the instances running the tests don't
recognise the iso_8859_16, and so disregard that encoding and try the
next suggested one - utf-8.

We expect all of our broadcast messages to be sent in utf-8, so let's
tell BeautifulSoup that explicitly.