Migration for "letter_branding_to_organisation" table
This adds a migration for a new table, letter_branding_to_organisation,
which matches the email_branding_to_organisation table.

The table has two columns, "organisation_id" and "letter_branding_id",
and the composite primary key on these means that we don't need to
create a separate unique constraint because uniqueness is already enforced.

This uses a `db.Table` instead of a `db.Model` (which is what
email_branding_to_organisation uses) in models.py. We don't need a model
class for the join table since we'll only be accessing the relationship
through an organisation or a letter branding object, and we don't need to
store extra data about the relationship.