Add endpoint to update letter branding pool
This adds an endpoint and new dao function to update the letter branding
pool for an organisation. The form that the data comes from allows multiple
letter brandings to be added at once, so the endpoint and dao function take
this into account by allowing the new brandings to be added in bulk.

As with other POST endpoints, we check the data using a JSON schema,
however the request from admin should always have the data in the
required format, so this acts more as documentation than a check that
might fail.