Add endpoint and dao function to remove letter branding from pool
This adds a new endpoint to remove a particular letter brand from the
letter branding pool for an organisation. As with the email branding
equivalent, there is a check that you're not removing the default and
that the brand you're trying to remove is in the pool. However, we do
also do the same checks in admin, so do not expect these errors to be
raised.