Give NHS letter branding the same ID in each environment
We gave the NHS email branding the same ID in each environment, which
simplifies the code to give it to organisations and to use it on forms.

This does the same for the NHS letter branding ID. The logo exists in
all environments. We delete the row in the letter_branding table from
all environments apart from production, then re-insert it to the other
environments with the ID that exists in production.