Add NHS letter branding to pools and org when creating an NHS org
When an organisation is created with one of the NHS org types, we now
set its letter branding to the NHS brand and add the NHS brand to its
branding pool.

We are already setting up NHS branding for services which are NHS
services, so this change won't really affect people's experience - it
just makes sure that the letter branding pools are set up with the right
data.