Migration to add NHS letter branding to letter branding pools
This migration adds the NHS letter branding to the branding pool for all
NHS organisations. There will be another migration later to add the
default branding to the pool for all organisations. However this
migration deals with the scenario where an NHS organisation does not
have NHS branding as its default - we still want to add it to the pool
in this case.