write the current alembic head to file when running migrations
this file isn't used in our code anywhere, however, if you have a
migration ready to go and someone else has merged a change, this'll mean
they also update this file, thus you'll get a merge conflict and will be
forced to rebase and fix the migration.

Even if you downgrade or aren't at head, this'll still write the most
recent version available.