Always add NHS branding to pools when updating an org
When we update an organisation to be one of the NHS org types we want to
add the NHS branding to the pool. We were already doing this, but had
missed one edge case - we weren't adding to the pool in cases where the
organisation already had a default branding set.

We now:
* Always add NHS branding to the pool if we're setting the org type to NHS
* Add branding to the pool if it's being set as the default for an org